Implement a lower bound for Arc<T> and Rc<T>
This crate has always given a lower bound of the actual heap memory usage:
we use `#[ignore_heap_size_of]` when not doing it is not practical.

This fix #37 with one of the ideas given in the discussion there:
return an accurate result when the reference count is 1
(and ownership is in fact not shared), or zero otherwise.

However, since APIs to access reference counts are not `#[stable]` yet
(see https://github.com/rust-lang/rust/issues/28356),
so always return zero when unstable APIs are not enabled.