Implement HeapSizeOf for raw pointers
These return 0 since we don't know if the pointer is valid or not. This
is a bit dubious, but is designed to support cases like LinkedHashMap
where you have some data structure like a HashMap used with raw pointers
in its type parameters. We want to be able to rely on the standard
HeapSizeOf implementation for HashMap, and then do the extra bookkeeping
of managing the space consumed by the raw pointers ourselves.