Accommodate change of empty vector representation in nightly libstd.
The empty sentinel value used to be 0x1; now it is the minimum
alignment of the contained type. To support both nightly and
stable versions of rustc, we check for any pointer that is equal
or less than the minimum alignment of the type that heap_size_of
receives. If the given pointer is larger, we pass it on to
the jemalloc heap-measuring API.